refactor: change file structure to (normal?)
