+++ src/gui/assets/logo.ico
