+++ src/gui/assets/logo.png
