+++ src/gui/style.qss
