+++ src/gui/translations/app_uk.ts
